Fixed deprecation warning about CURLOPT_PROGRESSFUNCTION.
Also now our background downloader uses int64 for tracking number of bytes per download.
Which might become useful if some FM crosses 2 GB line...

git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@11264 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ framework/FileSystem.h
@@ -76,8 +76,8 @@ typedef enum {
 typedef struct urlDownload_s {
 	idStr				url;
 	char				dlerror[ MAX_STRING_CHARS ];
-	int					dltotal;
-	int					dlnow;
+	int64				dltotal;
+	int64				dlnow;
 	int					dlstatus;
 	dlStatus_t			status;
 } urlDownload_t;
